# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rest Api - Laravel 11",
        "description": "Example of Rest Api by Laravel 11 platform",
        "contact": {
            "name": "ASt",
            "email": "alexandr.statut@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Dev API server"
        },
        {
            "url": "http://127.0.0.1:8001/api",
            "description": "Prod API server"
        }
    ],
    "paths": {
...
        "/v1/category": {
            "get": {
                "tags": [
                    "Категории"
                ],
                "summary": "Api: Список категорий",
                "description": "Список категорий",
                "operationId": "v1.category.getList",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Поиск по совпадению",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Test"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Статус",
                        "required": false,
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                1,
                                2
                            ],
                            "default": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "example": "1"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Записей на страницу",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "example": "10"
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Соротировка по полю",
                        "required": false,
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "title",
                                "status",
                                "created"
                            ],
                            "default": "title"
                        }
                    },
                    {
                        "name": "reverse",
                        "in": "query",
                        "description": "Реверс сортировки",
                        "required": false,
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "default": "asc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "title": "Список категорий",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CategoryResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server not available",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Категории"
                ],
                "summary": "Api: Добавить категорию",
                "description": "Добавить категорию",
                "operationId": "v1.category.store",
                "requestBody": {
                    "description": "Данные новой категории",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "title": "Заголовок",
                                        "type": "string",
                                        "example": "Test",
                                        "nullable": "false"
                                    },
                                    "status": {
                                        "title": "Статус",
                                        "type": "integer",
                                        "example": "1",
                                        "nullable": "true"
                                    }
                                },
                                "type": "object"
                            },
                            "examples": {
                                "Active category": {
                                    "summary": "Active category",
                                    "value": {
                                        "title": "Test 1",
                                        "status": 1
                                    }
                                },
                                "Blocked category": {
                                    "summary": "Blocked category",
                                    "value": {
                                        "title": "Test 2",
                                        "status": 2
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CategoryResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server not available",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CategoryResource": {
                "title": "Карточка категории",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Transform the resource into an array.",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "title": {
                        "title": "Заголовок",
                        "type": "string",
                        "example": "Test"
                    },
                    "status": {
                        "title": "Статус",
                        "type": "integer",
                        "example": "1"
                    },
                    "created": {
                        "title": "Создана",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-12-01 12:00:00"
                    },
                    "updated": {
                        "title": "Обновлена",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-12-01 12:00:00"
                    }
                },
                "type": "object"
            },
            "ErrorResource": {
                "title": "Ответ с ошибкой",
                "properties": {
                    "status": {
                        "title": "Статус",
                        "type": "boolean",
                        "format": "int64",
                        "example": "false"
                    },
                    "code": {
                        "title": "Код",
                        "type": "integer",
                        "format": "int64",
                        "example": "500"
                    },
                    "msg": {
                        "title": "Сообщение",
                        "type": "string",
                        "example": "Test"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Your token providing after auth in Api",
                "name": "Api Client",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}